New changes

+++ Final_Project/src/Business/Donation/Donation.java
@@ -9,34 +9,36 @@ package Business.Donation;
  *
  * @author mansi
  */
-public class Donation {
-    private int food_quantity;
-    private int cloth_quantity;
-    private int moneyDonation;
+public class Donation 
+{
+    private float food_quantity;
+    private float cloth_quantity;
+    private float moneyDonation;
     
 
    
-    public int getFood_quantity() {
+    public float getFood_quantity() 
+    {
         return food_quantity;
     }
 
-    public void setFood_quantity(int food_quantity) {
+    public void setFood_quantity(float food_quantity) {
         this.food_quantity = food_quantity;
     }
 
-    public int getCloth_quantity() {
+    public float getCloth_quantity() {
         return cloth_quantity;
     }
 
-    public void setCloth_quantity(int cloth_quantity) {
+    public void setCloth_quantity(float cloth_quantity) {
         this.cloth_quantity = cloth_quantity;
     }
 
-    public int getMoneyDonation() {
+    public float getMoneyDonation() {
         return moneyDonation;
     }
 
-    public void setMoneyDonation(int moneyDonation) {
+    public void setMoneyDonation(float moneyDonation) {
         this.moneyDonation = moneyDonation;
     }
